# ci-operator-prowgen: read/write jobs only once