This PR will serve as a cover letter for several others which may not seem
coherent on their own:

- https://github.com/openshift/ci-tools/pull/2307
- https://github.com/openshift/ci-tools/pull/2300
- https://github.com/openshift/ci-tools/pull/2301
- https://github.com/openshift/ci-tools/pull/2302
- https://github.com/openshift/ci-tools/pull/2303

The end result is a fairly non-intrusive change to `prowgen` in order to reduce
its total wall-clock execution time.  The main motivation is the fact that
`openshift/release` has got so big that merely performing a no-op job generation
as a pre-submit job in this repository takes longer than the execution of the
E2E tests.  Even on a developer machine, the total execution time is about a
minute or more.

The execution time is dominated by marshalling and unmarshalling YAML files,
mostly Prow jobs, and the accompanying garbage collection:

```
$ perf report -F overhead_children,symbol -g none \
>   | grep -m 3 -e yaml.Marshal -e yaml.Unmarshal -e runtime.gcDrain
    36.96%  [.] github.com/ghodss/yaml.Marshal
    29.50%  [.] github.com/ghodss/yaml.Unmarshal
    28.16%  [.] runtime.gcDrain
```

Although hard to optimize (Go and the `yaml` library do not seem to have any
concept of CPU or memory optimization), the problem can be alleviated by
reducing the number of times the same file is read/written:

```
$ strace -fe openat prowgen \
>   --from-dir ci-operator/config --to-dir ci-operator/jobs \
>   --registry ci-operator/step-registry \
>   |& perl -nE 'say $1 if /^\[pid \d+\] openat\(\w+, (.+?),/' | grep /ci-operator/ \
>   | sort | uniq -c | sort -n | tail -5
     14 "/home/bbguimaraes/src/release/ci-operator/jobs/openshift/odo/openshift-odo-main-presubmits.yaml"
     18 "/home/bbguimaraes/src/release/ci-operator/jobs/openshift-knative/serverless-operator/openshift-knative-serverless-operator-main-periodics.yaml"
     18 "/home/bbguimaraes/src/release/ci-operator/jobs/openshift-psap/ci-artifacts/openshift-psap-ci-artifacts-master-presubmits.yaml"
     20 "/home/bbguimaraes/src/release/ci-operator/jobs/openshift-knative/serverless-operator/openshift-knative-serverless-operator-main-presubmits.yaml"
     86 "/home/bbguimaraes/src/release/ci-operator/jobs/openshift/release/openshift-release-master-periodics.yaml"
```

Other optimizations include decoding files from `ci-operator/config` and
`ci-operator/jobs` in multiple threads and using the appropriate level of
concurrency for a given CPU, and there is room for improvement if time allows in
the future.

The relative and absolute execution times of relevant changes operating on
`openshift/release@fbdba5c086` are (not in commit order, all pass integration
and no-op generation tests):

```
total time(s) |   from start    |    from prev.   | revision
------------------------------------------------------------
 37.054826892 | -  0.00, * 1.00 | -  0.00, * 1.00 | eaa6ae83b [master]
 24.688385406 | - 12.37, * 0.67 | - 12.37, * 0.67 | 7c97f58fb ci-operator-prowgen: accumulate input
  6.863990980 | - 30.19, * 0.19 | - 17.82, * 0.28 | 23410c1d2 ci-operator-prowgen: read/write jobs only once
  6.651639983 | - 30.40, * 0.18 | -  0.21, * 0.97 | 496c644f1 ci-operator-prowgen: use `nproc` threads
  5.556544602 | - 31.50, * 0.15 | -  1.10, * 0.84 | a3cfe2b1e pkg/config: load in parallel
  5.539440110 | - 31.52, * 0.15 | -  0.02, * 1.00 | 67f0f904f pkg/jobconfig: load in parallel
```

The optimal number of reads/writes is achieved:

```
$ strace -fe openat prowgen \
>   --from-dir ci-operator/config --to-dir ci-operator/jobs \
>   --registry ci-operator/step-registry \
>   |& perl -nE 'say $1 if /^\[pid \d+\] openat\(\w+, (.+?),/' | grep /ci-operator/ \
>   | sort | uniq -c \
>   | awk '{print$1}' | sort -n | uniq -c
   7942 1
   9675 2
```

The bottlenecks remain the same, but are much less pronounced and the final
execution time should be satisfactory for the current size of the repository.

```
$ perf report --fields overhead_children,symbol --call-graph none \
>   | grep -m 3 -e yaml.Marshal -e yaml.Unmarshal -e runtime.gcDrain
    42.62%  [.] github.com/ghodss/yaml.Marshal
    38.54%  [.] github.com/ghodss/yaml.Unmarshal
    12.93%  [.] runtime.gcDrain
```